@@ -1,14 +1,13 @@
-const functions = require('firebase-functions');
-const admin = require('firebase-admin');
-const axios = require('axios');
-const cors = require('cors')({origin: true});
-const express = require('express');
-const { FUNCTIONS_CONFIG } = require("./FirebaseEnvironment.js");
-const { updateUserWithStrava } = require("./updateUserWithStrava");
-const { refreshToken } = require("./refreshToken");
+import * as functions from 'firebase-functions';
+import { FUNCTIONS_CONFIG } from "./FirebaseEnvironment";
+import axios from 'axios';
+// import cors from 'cors';
+import * as express from 'express';
+import { updateUserWithStrava } from "./updateUserWithStrava";
+import { refreshToken } from "./refreshToken";
 
 const app = express();
-app.get('/stravaauth', async (req, res) => {
+app.get('/stravaauth', async (req:any, res:any) => {
     console.log(`called stravaSendAuthorizationRequest with environment ${req}`);
     const redirectURL = req.query.redirect_uri ? req.query.redirect_uri : FUNCTIONS_CONFIG.strava.redirect_uri;
     console.log(`redirectURL: ${redirectURL}`)
@@ -38,7 +37,7 @@ app.get('/stravaauth', async (req, res) => {
         res.end(200);
     } catch (err) {
         console.error(`Error redirecting to URI: ${err}`);
-        res.redirect(`${redirect_uri}?error=BadRequest`);
+        res.redirect(`${redirectURL}?error=BadRequest`);
         return res.status(401).json(`Error Caught in OAuth server request: ${err}`);
     }
 
@@ -46,17 +45,17 @@ app.get('/stravaauth', async (req, res) => {
 exports.oauth = functions.https.onRequest(app);
 
 // This gets the token after the user has authorized the app to use strava
-exports.stravaGetToken = functions.https.onCall((req, res) => {
+exports.stravaGetToken = functions.https.onCall((req:any, context:any) => {
     return new Promise((resolve, reject) => {
         console.log(`called stravaGetToken with request`);
         console.log(req);
-        let code = undefined;
-        let uid = undefined;
+        let code:string = "";
+        let uid:string = "";
         if (req && req.code && req.uid) {
             code = req.code;
             uid = req.uid;
         } else {
-            return reject(`Error in stravaGetToken - invalid parm - must provide code and uid`);
+            reject(`Error in stravaGetToken - invalid parm - must provide code and uid`);
         }
 
         const params = {
@@ -74,32 +73,32 @@ exports.stravaGetToken = functions.https.onCall((req, res) => {
             ;
 
         console.log(`URIRequest: ${URIRequest}`);
-        axios.post(URIRequest).then((res) => {
+        axios.post(URIRequest).then((response:any) => {
             // console.log("Successfully sent strava token request.  response info");
 
             // MUST break up response since FB functions can not resolve this complex
             // res object as it has circular reference.  It causes an error;
             // Unhandled error RangeError: Maximum call stack size exceeded
             const clientResponse = {
-                refresh_token: res.data.refresh_token,
-                access_token: res.data.access_token,
-                expires_at: res.data.expires_at,
-                expires_in: res.data.expires_in,
-                athlete: res.data.athlete,
-                athleteId: res.data.athlete.id,
+                refresh_token: response.data.refresh_token,
+                access_token: response.data.access_token,
+                expires_at: response.data.expires_at,
+                expires_in: response.data.expires_in,
+                athlete: response.data.athlete,
+                athleteId: response.data.athlete.id,
             }
 
             // NOW, save the strava info to the user account, maybe do in client
             // Must Save access_toke, refresh_token, expires_at, accepted_scopes
-            updateUserWithStrava(uid, clientResponse).then (() => {
-                return resolve(clientResponse);
-            }).catch((err) => {
+            updateUserWithStrava(uid, clientResponse, false).then (() => {
+                resolve(clientResponse);
+            }).catch((err:any) => {
                 console.error(`FB Func: stravaGetToken in updateUserWithStrava -- Error : ${err}`);
-                return reject(`FB Func: stravaGetToken in updateUserWithStrava -- Error : ${err}`); 
+                reject(`FB Func: stravaGetToken in updateUserWithStrava -- Error : ${err}`); 
             });
-        }).catch((err) => {
-            console.error(`FB Func: stravaGetToken -- Error : ${err}`);
-            return reject(`FB Func: stravaGetToken -- Error : ${err}`); 
+        }).catch((err2:any) => {
+            console.error(`FB Func: stravaGetToken -- Error : ${err2}`);
+            reject(`FB Func: stravaGetToken -- Error : ${err2}`); 
         });
     });
 });
@@ -107,17 +106,17 @@ exports.stravaGetToken = functions.https.onCall((req, res) => {
 // refactored the code into separate file so I could share front and backend calling
 exports.stravaRefreshToken = functions.https.onCall(refreshToken);
 
-exports.stravaDeauthorize = functions.https.onCall((req, res) => {
+exports.stravaDeauthorize = functions.https.onCall((req:any, context:any) => {
     return new Promise((resolve, reject) => {
         console.log(`called stravaDeauthorize with request`);
         console.log(req);
-        let access_token = undefined;
-        let uid = undefined;
+        let access_token:string = "";
+        let uid:string = "";
         if (req && req.access_token && req.uid) {
             access_token = req.access_token;
             uid = req.uid;
         } else {
-            return reject(`Error in stravaDeauthorize - invalid parm - must provide uid and access_token`);
+            reject(`Error in stravaDeauthorize - invalid parm - must provide uid and access_token`);
         }
 
         console.log(`stravaDeauthorize access_token${access_token}, uid:${uid}`);
@@ -139,33 +138,32 @@ exports.stravaDeauthorize = functions.https.onCall((req, res) => {
                 headers: {
                     'Authorization': `Bearer ${access_token}`
                 }
-            }).then((res) => {
-            console.log("Successfully sent stravaDeauthorize refresh request.");
+            }).then((response: any) => {
+                console.log("Successfully sent stravaDeauthorize refresh request.");
 
-            // MJUST break up response since FB functions can not resolve this complex
-            // res object as it has circular reference.  It causes an error;
-            // Unhandled error RangeError: Maximum call stack size exceeded
-            const clientResponse = {
-                refresh_token: res.data.refresh_token,
-            }
+                // MJUST break up response since FB functions can not resolve this complex
+                // res object as it has circular reference.  It causes an error;
+                // Unhandled error RangeError: Maximum call stack size exceeded
+                const clientResponse = {
+                    refresh_token: response.data.refresh_token,
+                }
 
-            // NOW, save the strava info to the user account, maybe do in client
-            // Must Save access_toke, refresh_token, expires_at, accepted_scopes
-            const deauthorize = true;
-            updateUserWithStrava(uid, clientResponse, deauthorize).then (() => {
-                resolve(clientResponse);
-            }).catch((err) => {
-                console.error(`stravaDeauthorize->updateUserWithStrava -- ${err}`);
-                reject(`stravaDeauthorize->updateUserWithStrava -- ${err}`); 
-            });
-        }).catch((err) => {
-            console.error(`stravaDeauthorize failed -- ${err}`);
-            reject(`stravaDeauthorize failed -- ${err}`); 
+                // NOW, save the strava info to the user account, maybe do in client
+                // Must Save access_toke, refresh_token, expires_at, accepted_scopes
+                updateUserWithStrava(uid, clientResponse, true).then (() => {
+                    resolve(clientResponse);
+                }).catch((err:any) => {
+                    console.error(`stravaDeauthorize->updateUserWithStrava -- ${err}`);
+                    reject(`stravaDeauthorize->updateUserWithStrava -- ${err}`); 
+                });
+        }).catch((err2: any) => {
+            console.error(`stravaDeauthorize failed -- ${err2}`);
+            reject(`stravaDeauthorize failed -- ${err2}`); 
         });
     });
 });
 
-exports.stravaGetActivities = functions.https.onCall((req, context) => {
+exports.stravaGetActivities = functions.https.onCall((req:any, context:any) => {
     // curl -X GET "https://www.strava.com/api/v3/athlete/activities?after=1546318800&page=1&per_page=30"
     // -H  "accept: application/json" -H  "authorization: Bearer access_token"
     return new Promise((resolve, reject) => {
@@ -179,7 +177,7 @@ exports.stravaGetActivities = functions.https.onCall((req, context) => {
             access_token = req.access_token;
             uid = req.uid;
         } else {
-            return reject(`Error in stravaGetActivities - invalid parm - must provide uid and access_token`);
+            reject(`Error in stravaGetActivities - invalid parm - must provide uid and access_token`);
         }
         
         if (req && req.dateAfter) {
@@ -211,7 +209,7 @@ exports.stravaGetActivities = functions.https.onCall((req, context) => {
                     'authorization': `Bearer ${access_token}`,
                     'accept': "application/json"
                 }
-            }).then((res) => {
+            }).then((res:any) => {
             console.log("Successfully sent stravaGetActivities request.");
 
             // MUST break up response since FB functions can not resolve this complex
@@ -222,7 +220,7 @@ exports.stravaGetActivities = functions.https.onCall((req, context) => {
             // console.log("Response data (res.data)");
             // console.log(res.data);
             resolve(res.data);
-        }).catch((err) => {
+        }).catch((err:any) => {
             console.error(`stravaGetActivities failed -- ${err}`);
             reject(`stravaGetActivities failed -- ${err}`); 
         });
