@@ -2,7 +2,7 @@ import * as functions from 'firebase-functions';
 import * as admin from 'firebase-admin';
 
 exports.authDeleteUser = functions.https.onCall((req, context) => {
-    let uid = req.uid;
+    const uid = req.uid;
     // delete the authUser
     // MUST RETURN A PROMISE!
     return new Promise((resolve, reject) => {
@@ -26,16 +26,15 @@ exports.authCreateUser = functions.https.onCall((user, res) => {
              
     // Generate random password if no password exists (meaning its created by someone else)
     if (!user.password || user.password === null || user.password === "") {
-        let randomPassword = Math.random().toString(36).slice(-8);
+        const randomPassword = Math.random().toString(36).slice(-8);
         user.password = randomPassword;
     }
     // First, see if user already exists in auth
     // MUST return promise
     return new Promise((resolve, reject) => {
-        admin.auth().getUserByEmail(user.email).then((user) => {
+        admin.auth().getUserByEmail(user.email).then((newAuthUser) => {
             // See the UserRecord reference doc for the contents of userRecord.
-            let authUser = {};
-            authUser = user;
+            const authUser = newAuthUser;
             console.log(`Successfully fetched user data: ${JSON.stringify(authUser)}`);
             resolve(authUser);
         }).catch((err) => {
@@ -50,10 +49,10 @@ exports.authCreateUser = functions.https.onCall((user, res) => {
             }).then((authUser) => {
                 console.log(`Successfully added auth user: ${JSON.stringify(authUser)}`);
                 resolve(authUser);
-            }).catch((err) => {
+            }).catch((err2) => {
                 //throw Error(`Error creating auth user in authCreateUser error: ${err}`);
-                console.error(`Error creating auth user in authCreateUser error: ${err}`);
-                reject(`Error creating auth user in authCreateUser error: ${err}`);
+                console.error(`Error creating auth user in authCreateUser error: ${err2}`);
+                reject(`Error creating auth user in authCreateUser error: ${err2}`);
             });
         });  // get
     });
